[#34] refactor(profile): decompose DetailsTab into modular ViewMode and EditMode components
